requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioredis==1.3.1
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.6
decorator==5.1.1
distlib==0.3.8
Django==5.1.4
django-channels==0.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.5.0
graphviz==0.20.3
h11==0.14.0
h2==4.1.0
hiredis==3.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jwt==1.3.1
kiwisolver==1.4.5
kombu==5.4.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==2.0.0
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
priority==2.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyparsing==3.1.2
pytest==8.3.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==307
pyzmq==26.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
upstash-redis==1.2.0
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
virtualenv==20.26.3
vunit_hdl==4.7.0
wcwidth==0.2.13
websocket-client==1.8.0
whitenoise==6.8.2
wsl==0.12
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
